html/syntax/serializing-html-fragments/processing-instructions.html

<!DOCTYPE html>
<title>ProcessingInstruction serialization in HTML</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  const pi = document.createProcessingInstruction("target", "data");
  const div = document.createElement("div");
  div.appendChild(pi);
  assert_equals(div.innerHTML, "<?target data>");
}, "PI created by document.createProcessingInstruction()");

test(function() {
  const parser = new DOMParser();
  const xmlDoc = parser.parseFromString("<root><?target data?></root>", "text/xml");
  const pi = xmlDoc.documentElement.firstChild;
  const div = document.createElement("div");
  div.appendChild(pi);
  assert_equals(div.innerHTML, "<?target data>");
}, "PI adopted from XML document");
</script>
